organizando codigo.
