+++ README.md
@@ -2,3 +2,8 @@ TwitterSentiment
 ================
 
 Analyze the sentiment of tweets using Java
+
+Required
+=============
+lingpipe < http://alias-i.com/lingpipe/web/download.html >
+twitter4j < http://twitter4j.org/en/index.html >
