--- train.java
@@ -1,23 +1,0 @@
-void train() throws IOException, ClassNotFoundException {
-	File trainDir;
-	String[] categories;
-	LMClassifier class;
-	trainDir = new File("trainDirectory");
-	categories = trainDir.list();
-	int nGram = 7; //the nGram level, any value between 7 and 12 works
-	class= DynamicLMClassifier.createNGramProcess(mCategories, nGram);
-
-	for (int i = 0; i < categories.length; ++i) {
-		String category = categories[i];
-		Classification classification = new Classification(category);
-		File file = new File(trainDir, categories[i]);
-		File[] trainFiles = file.listFiles();
-		for (int j = 0; j < trainFiles.length; ++j) {
-			File trainFile = trainFiles[j];
-			String review = Files.readFromFile(trainFile, "ISO-8859-1");
-			Classified classified = new Classified(review, classification);
-			((ObjectHandler>) class).handle(classified);  
-		}
- 	}
-	AbstractExternalizable.compileTo((Compilable) class, new File("classifier.txt"));
-}
